refactor(ia32): Refactor segment descriptor module
Split segment descriptor module with two new submodules, one for every
sub structure. This will allow shorter and cleaner code files.
